Разделяет код на модули
