+++ js/data.js
@@ -1,0 +1,171 @@
+import { getRandomPositiveInteger, getRandomGeographicCoordinate, getRandomArrayElement } from './util.js';
+
+const TITLES = [
+  '1-комн. квартира, 45 м²',
+  'Круглосуточное заселение!',
+  'Квартира Свободна. ТОП! Звони!',
+  'Объявление проверено. Гарантия.',
+  'Лучшие апартаменты',
+  'Дом с бассейном и 8 спальнями',
+  'Лучшее предложение!',
+  'Коттедж для праздников',
+  '2-этажный дом, 550 м²',
+  'Коттедж с бассейном',
+];
+
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+const CHECKS = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const DESCRIPTIONS = [
+  `Просторные апартаменты в центре Москвы!
+  Квартира порадует Вас дизайнерским интерьером, широкоформатными окнами и доступностью к главным достопримечательностям столицы.`,
+  `Светлая  просторная квартира с дизайнерским интерьером в центре столицы. Вас не потревожит шум с улиц, так как окна квартиры выходят во двор. Уютные спальные комнаты обеспечат Вам комфортный отдых и крепкий сон. Просторная гостиная с обеденной зоной и уютным диваном позволит Вам душевно проводить вечера после насыщенного впечатлениями и эмоциями дня.
+  Кухня оснащена всей бытовой техникой: посудомоечной машиной, холодильником, микроволновой печью, духовым шкафом и всей необходимой посудой. Во дворе расположена детская площадка. Мы предоставляем парковку на территории двора (за дополнительную плату)`,
+  `просторные и светлые трехкомнатные апартаменты с современным дизайном и качественным ремонтом. Наш дизайнер создал теплую и созидающую атмосферу в центре мегаполиса.
+  В квартире две спальные комнаты и гостиная. В гостиной удобный раскладной двуспальный диван, плазменный TV с функцией SMART TV, журнальный столик с уютным креслом. Из гостиной Вы попадете на открытый балкон, где, сидя за столиком, с чашкой ароматного кофе, можете насладиться видом московских двориков.
+  В спальных комнатах удобные двуспальные кровати. `,
+  `Кухня полностью оснащена бытовой техникой: варочной панелью, духовым шкафом, посудомоечной машиной, эл. чайником, микроволновой печью, блендером, тостером, а также всей необходимой посудой. в апартаментах два санузла.
+  В стоимость проживания входит: высокоскоростной Wi-Fi (100 МБ/сек), кабельное телевидение, чай, кофе, комплекты постельного белья и полотенец , фен, утюг, гигиенические принадлежности, тапочки. Во время проживания проводится регулярная уборка апартаментов.
+  Обратите внимание, данная квартира не сдается под мероприятия, в том числе под Дни Рождения по ряду определенных причин.`,
+  `Роскошная трешка с дизайнерским интерьером! В вашем распоряжении две спальные комнаты, гостиная, просторная кухня.
+  Спальные комнаты оснащены двуспальными кроватями, вместительными шкафами, комодом, прикроватными тумбами и светильниками. Хорошо выспаться Вам помогут шторы на окнах из плотной ткани. Благодаря глубоким сочным цветам в интерьере, в сочетании с дорогой итальянской мебелью из массива дуба вы ощутите особую атмосферу уюта и покоя в центре города.
+  Для Вашего комфорта есть вся необходимая посуда и бытовая техника (телевизор с плоским экраном, микроволновая печь, электрический чайник, газовая плита, фен, утюг). В квартире возможно размещение до 6 человек.`,
+  `Апартаменты находятся на 19 этаже, откуда открывается изумительный панорамный вид!
+  Стильная, после капитального ремонта, квартира обеспечит Вам хороший отдых и незабываемые впечатления. Просторная гостиная площадью 27  кв. м. оснащена двуспальным диваном, обеденной зоной.  В спальне в Вашем распоряжении двуспальная кровать, вместительный шкаф.  В каждой комнате установлены  ЖК телевизоры.  Кухня полностью оборудована бытовой техникой: холодильником, микроволновой печью, посудомоечной машиной, электрическим чайником, варочной панелью, духовым шкафом, фильтром для воды.
+  В стоимость проживания входит: высокоскоростной  Wi-Fi (100 МБ/сек), кабельное телевидение, чай, кофе,  комплекты постельного белья и полотенец , фен, утюг, гигиенические принадлежности, тапочки.
+  Во время проживания проводится регулярная уборка апартаментов. Фото соответствует действительности. Прямая аренда без посредников и дополнительных комиссий. Для командировочных предоставляем отчетные документы. Поздний заезд после 00:00 оплачивается дополнительно.
+  Обратите внимание, данная квартира не сдается под мероприятия в т.ч. под Дни Рождения по ряду определенных причин.`,
+  `Площадь дома 600 кв.м , на участке 8 соток.
+
+  1-этаж: Гостиная с большим обеденным столом на 20 человек, камин, плазменный телевизор .
+  Одна спальная комната с двуспальной кроватью, сан. узел, джакузи, биде, стиральная машина.
+  Зона кухни: холoдильник , газовая плита с духовкой, микроволновая печь, посудомоечная машина, обеденный стол, посуда, теплые полы.
+  2-этаж: 4 Спальни.
+  3-этаж: 2 Спальни.
+
+  Спа-комплекс включает огромный плавательный бассейн с обогревом 20*3 м. Баня с комнатой отдыха, электрическая плита, сан узел.
+  Бильярдный зал (русский бильярд) +барная стойка, диван, комната отдыха с двухспальной кроватью.
+
+  Всего в доме 8 спальных комнат ( в каждой комнате двуспальная кровать)+ диваны, 5 сан узлов + на каждом этаже свой сан узел с душем.
+
+  На участке:  мангал, барбекю, беседка, хвойные деревья.`,
+  `Коттедж общей площадью 1000 кв.м. выполнен в современном стиле loft. В доме 14 комфортабельных спален, 2 из них двухкомнатные , 14 с/у. Одна из спален представляет собой апартаменты премиум класса для молодожёнов и подходит для проведения свадебных фотосессий.
+  В спальнях коттеджа можно разместить до 45 гостей.
+  Уникальной особенностью коттеджа является просторный банкетный зал в стиле loft, оборудованный камином.
+  Вид из панорамных окон зала открывается на живописный водоем.
+  Банкетный зал рассчитан на проведение банкетов до 80 и фуршетов до 120 персон.
+
+  На цокольном этаже расположен бар, домашний кинотеатр и ночной клуб с повышенной шумоизоляцией.
+  Ночной клуб оборудован светотехникой и акустической системой премиум класса мощностью 2 кВт.Клуб подходит для проведения концертов, в нем располагается сцена и гримерка.Дополнительно можно арендовать профессиональную караоке систему.
+
+  На огромном участке в 40 соток расположена беседка, которая в теплое время года используется как летний ресторан с возможностью проведения банкетов до 150 персон.
+  Рядом с водоемом оборудована площадка для выездной регистрации.На участке выполнен ландшафтный дизайн.Также в доме имеется большой бассейн с аквапарком, русский бильярд и баня на дровах.`,
+  `У нас вы сможете отпраздновать свадьбы, дни рождения как взрослых, так и детей, юбилеи, провести тимбилдинги и тематические вечеринки и, наконец, просто расслабиться после трудовой недели.
+  На первом этаже расположены общая кухня с большим обеденным столом на человек, посудой и необходимым оборудованием, кают-компания с суперсовременным караоке на 40 000 песен, TV и настоящим камином. Так же здесь есть одна спальня, туалет и ванная комната с джакузи.
+  На втором этаже находится шесть спален, душ и туалет, фойе с библиотекой. Во всех комнатах оборудованы полы с подогревом.`,
+  `Наш дом идеально подходит для длительного семейного проживания, проживания сотрудников компании, празднования свадеб, банкетов и других праздников.
+  Также к Вашим услугам организация банкетов, свадеб под ключ, кухня с прекрасным выбором блюд.
+  Звоните, уточняйте стоимость на интересующие Вас даты и приезжайте на просмотр!
+  - бaнкетный зaл
+  - обоpудованная кухня
+  - три санузла
+  - 6 изoлирoвaнных cпалeн
+  - 20 cпальныx мecт
+  Цена указана за будний день до 12человек,выходные и праздничные дни обговариваются по телефону,цена зависит от сезонности,состава компаний и мероприятия!
+  Уличная инфpaструктуpа:
+  - мангал для шашлыка
+  - удобный, асфальтированный, освещенный подъезд до самого коттеджа
+  - магазин самообслуживания в 300м со всеми необходимыми товарами.`,
+];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+const GEOGRAPHIC_COORDINATES = {
+  LAT_MIN: 35.65,
+  LAT_MAX: 35.7,
+  LNG_MIN: 139.7,
+  LNG_MAX: 139.8,
+};
+
+const SIMILAR_OFFERS_COUNT = 10;
+const DECIMALS_OF_GEOGRAPHIC_COORDINATES = 5;
+let avatarCounter = SIMILAR_OFFERS_COUNT + 1;
+
+function createFeatures(length) {
+  const features = FEATURES.slice();
+  const numberElementsToBeRemoved = features.length - length;
+
+  for (let i = 0; i < numberElementsToBeRemoved; i++) {
+    features.splice(getRandomPositiveInteger(0, features.length - 1), 1);
+  }
+
+  return features;
+}
+
+function createPhotos(length) {
+  const photos = [];
+
+  for (let i = 0; i < length; i++) {
+    photos.push(PHOTOS[getRandomPositiveInteger(0, PHOTOS.length - 1)]);
+  }
+
+  return photos;
+}
+
+function createOffer() {
+  avatarCounter--;
+
+  const locationLat = getRandomGeographicCoordinate(GEOGRAPHIC_COORDINATES.LAT_MIN, GEOGRAPHIC_COORDINATES.LAT_MAX, DECIMALS_OF_GEOGRAPHIC_COORDINATES);
+  const locationLng = getRandomGeographicCoordinate(GEOGRAPHIC_COORDINATES.LNG_MIN, GEOGRAPHIC_COORDINATES.LNG_MAX, DECIMALS_OF_GEOGRAPHIC_COORDINATES);
+
+  return {
+    author: {
+      avatar: avatarCounter > 9 ? `img/avatars/user${avatarCounter}.png` : `img/avatars/user0${avatarCounter}.png`,
+    },
+    offer: {
+      title: getRandomArrayElement(TITLES),
+      address: `${locationLat}, ${locationLng}`,
+      price: getRandomPositiveInteger(1, 99999),
+      type: getRandomArrayElement(TYPES),
+      rooms: getRandomPositiveInteger(1, 10),
+      guests: getRandomPositiveInteger(1, 15),
+      checkin: getRandomArrayElement(CHECKS),
+      checkout: getRandomArrayElement(CHECKS),
+      features: createFeatures(getRandomPositiveInteger(1, FEATURES.length)),
+      description: getRandomArrayElement(DESCRIPTIONS),
+      photos: createPhotos(getRandomPositiveInteger(1, 10)),
+    },
+    location: {
+      lat: locationLat,
+      lng: locationLng,
+    },
+  };
+}
+
+const createOffers = () => Array.from({ length: SIMILAR_OFFERS_COUNT }, createOffer);
+
+export { createOffers };
